[#198] Refactor: global, infra 패키지 정리
