+++ src/main/java/com/vt/valuetogether/infra/s3/S3Config.java
@@ -1,4 +1,4 @@
-package com.vt.valuetogether.global.s3;
+package com.vt.valuetogether.infra.s3;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
